requirements.txt
amqp==1.4.9
anyjson==0.3.3
backoff==1.6.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto==2.49.0
cachetools==4.0.0
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.9
django-analytical==2.5.0
django-celery==3.3.1
django-crispy-forms==1.8.1
django-google-analytics-app==4.4.1
django-heroku==0.3.1
django-modelcluster==5.0.1
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
kombu==3.0.37
monotonic==1.5
Pillow==6.2.2
posthog==1.0.10
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
structlog==20.1.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7.1
Wand==0.5.8
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
